faire exception

+++ Integgreat.Application/Services/Impl/TaskService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Integgreat.Application.DTOs.Task;
+using Integgreat.Application.Exceptions;
 using Integgreat.Domain.Entities;
 using Integgreat.Domain.Interfaces;
 using TaskStatus = Integgreat.Domain.Enums.TaskStatus;
@@ -40,7 +41,7 @@ public class TaskService : ITaskService
     public async Task<TaskResponseDto> UpdateStatusAsync(int id, TaskStatus status)
     {
         var task = await _taskRepository.GetByIdAsync(id);
-        if (task == null) throw new Exception("Task not found");
+        if (task == null) throw new NotFoundException("Task not found");
         task.Status = status;
         await _taskRepository.UpdateAsync(task);
         return _mapper.Map<TaskResponseDto>(task);
@@ -48,6 +49,8 @@ public class TaskService : ITaskService
 
     public async Task DeleteAsync(int id)
     {
+        var task = await _taskRepository.GetByIdAsync(id);
+        if (task == null) throw new NotFoundException("Task not found");
         await _taskRepository.DeleteAsync(id);
     }
 }
